rename ConfigUtils to PropertiesUtils; expose more property fetching methods on ConfigurationUtils
