+++ dubbo-container/dubbo-container-spring/src/main/java/org/apache/dubbo/container/spring/SpringContainer.java
@@ -16,9 +16,9 @@
  */
 package org.apache.dubbo.container.spring;
 
+import org.apache.dubbo.common.config.ConfigurationUtils;
 import org.apache.dubbo.common.logger.Logger;
 import org.apache.dubbo.common.logger.LoggerFactory;
-import org.apache.dubbo.common.utils.ConfigUtils;
 import org.apache.dubbo.common.utils.StringUtils;
 import org.apache.dubbo.container.Container;
 
@@ -42,7 +42,7 @@ public class SpringContainer implements Container {
 
     @Override
     public void start() {
-        String configPath = ConfigUtils.getProperty(SPRING_CONFIG);
+        String configPath = ConfigurationUtils.getProperty(SPRING_CONFIG);
         if (StringUtils.isEmpty(configPath)) {
             configPath = DEFAULT_SPRING_CONFIG;
         }
